# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Him For You discovery and membership API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /registration/profile:
    {
      "put":
        {
          "operationId": "updateMyRegistration",
          "tags": ["registration"],
          "requestBody":
            {
              "required": true,
              "content":
                {
                  "application/json":
                    {
                      "schema":
                        { "$ref": "#/components/schemas/RegistrationRecord" },
                    },
                },
            },
          "responses":
            {
              "200":
                {
                  "description": "Updated private profile",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "$ref": "#/components/schemas/RegistrationRecord",
                            },
                        },
                    },
                },
            },
        },
    }
...
  /registrations:
    {
      "post":
        {
          "operationId": "registerProfile",
          "tags": ["registration"],
          "responses":
            {
              "200":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "$ref": "#/components/schemas/RegistrationRecord",
                            },
                        },
                    },
                },
            },
          "requestBody":
            {
              "required": true,
              "content":
                {
                  "application/json":
                    {
                      "schema":
                        { "$ref": "#/components/schemas/RegistrationInput" },
                    },
                },
            },
        },
    }
...
  /profiles:
    get:
      operationId: listProfiles
      tags: [discovery]
      summary: Browse approved male profiles
      parameters:
        - $ref: "#/components/parameters/City"
        - $ref: "#/components/parameters/LookingFor"
        - $ref: "#/components/parameters/MinAge"
        - $ref: "#/components/parameters/MaxAge"
        - $ref: "#/components/parameters/Premium"
        - $ref: "#/components/parameters/Verified"
        - $ref: "#/components/parameters/Active"
        - $ref: "#/components/parameters/Sort"
      responses:
        "200":
          description: Matching profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Profile"
  /profiles/featured:
    get:
      operationId: getFeaturedProfiles
      tags: [discovery]
      summary: Get featured profiles for the homepage
      parameters:
        - $ref: "#/components/parameters/City"
      responses:
        "200":
          description: Featured profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Profile"
  /profiles/{slug}:
    get:
      operationId: getProfile
      tags: [discovery]
      summary: Get a public profile by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileDetail"
        "404":
          description: Profile not found
  /profiles/{id}/interest:
    post:
      operationId: sendInterest
      tags: [engagement]
      summary: Send interest to a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterestInput"
      responses:
        "201":
          description: Interest recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EngagementResponse"
  /profiles/{id}/favorite:
    post:
      operationId: toggleFavorite
      tags: [engagement]
      summary: Toggle a profile favourite
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Favourite state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EngagementResponse"
...
components:
  parameters:
    City:
      name: city
      in: query
      schema:
        type: string
    LookingFor:
      name: lookingFor
      in: query
      schema:
        type: string
    MinAge:
      name: minAge
      in: query
      schema:
        type: number
        minimum: 18
    MaxAge:
      name: maxAge
      in: query
      schema:
        type: number
        maximum: 100
    Premium:
      name: premium
      in: query
      schema:
        type: boolean
    Verified:
      name: verified
      in: query
      schema:
        type: boolean
    Active:
      name: active
      in: query
      schema:
        type: boolean
    Sort:
      name: sort
      in: query
      schema:
        type: string
        enum: [featured, active, newest, age]
        default: featured
  schemas:
...
    Profile:
      type: object
      properties:
        id:
          type: number
        slug:
          type: string
        displayName:
          type: string
        age:
          type: number
        city:
          type: string
        citySlug:
          type: string
        headline:
          type: string
        imageUrl:
          type: string
        interests:
          type: array
          items:
            type: string
        lookingFor:
          type: array
          items:
            type: string
        isVerified:
          type: boolean
        isPremium:
          type: boolean
        isFeatured:
          type: boolean
        isOnline:
          type: boolean
        lastActive:
          type: string
        favouriteCount:
          type: number
      required:
        [
          id,
          slug,
          displayName,
          age,
          city,
          citySlug,
          headline,
          imageUrl,
          interests,
          lookingFor,
          isVerified,
          isPremium,
          isFeatured,
          isOnline,
          lastActive,
          favouriteCount,
        ]
    ProfileDetail:
      allOf:
        - $ref: "#/components/schemas/Profile"
        - type: object
          properties:
            bio:
              type: string
            gallery:
              type: array
              items:
                type: string
            availability:
              type: string
            responseTime:
              type: string
          required: [bio, gallery, availability, responseTime]
...
    InterestInput:
      type: object
      required: [contactType, contact]
      properties:
        contactType:
          type: string
          enum: [telegram, whatsapp]
        contact:
          type: string
          minLength: 1
          maxLength: 64
        note:
          type: string
          maxLength: 500
    EngagementResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        isFavourite:
          type: boolean
      required: [success, message]
...
    RegistrationPhoto:
      {
        "type": "object",
        "properties":
          {
            "dataUrl":
              { "type": "string", "minLength": 1, "maxLength": 1500000 },
            "category":
              {
                "type": "string",
                "enum": ["profile", "additional", "gallery"],
              },
          },
        "required": ["dataUrl", "category"],
      }
    RegistrationInput:
      {
        "type": "object",
        "properties":
          {
            "displayName":
              { "type": "string", "minLength": 1, "maxLength": 80 },
            "dateOfBirth":
              { "type": "string", "minLength": 1, "maxLength": 10 },
            "email": { "type": "string", "minLength": 1, "maxLength": 254 },
            "mobile": { "type": "string", "minLength": 1, "maxLength": 25 },
            "password": { "type": "string", "minLength": 8, "maxLength": 128 },
            "confirmPassword":
              { "type": "string", "minLength": 1, "maxLength": 128 },
            "headline": { "type": "string", "minLength": 0, "maxLength": 120 },
            "about": { "type": "string", "minLength": 0, "maxLength": 2000 },
            "country": { "type": "string", "minLength": 0, "maxLength": 80 },
            "state": { "type": "string", "minLength": 0, "maxLength": 80 },
            "city": { "type": "string", "minLength": 0, "maxLength": 80 },
            "area": { "type": "string", "minLength": 0, "maxLength": 100 },
            "age": { "type": "number", "minimum": 18, "maximum": 100 },
            "pinCode": { "type": "string", "minLength": 0, "maxLength": 12 },
            "interests":
              {
                "type": "array",
                "minItems": 0,
                "maxItems": 20,
                "uniqueItems": true,
                "items": { "type": "string", "minLength": 1, "maxLength": 80 },
              },
            "languages":
              {
                "type": "array",
                "minItems": 0,
                "maxItems": 20,
                "uniqueItems": true,
                "items": { "type": "string", "minLength": 1, "maxLength": 80 },
              },
            "preferences":
              {
                "type": "array",
                "minItems": 0,
                "maxItems": 20,
                "uniqueItems": true,
                "items": { "type": "string", "minLength": 1, "maxLength": 80 },
              },
            "availability":
              {
                "type": "array",
                "minItems": 0,
                "maxItems": 20,
                "uniqueItems": true,
                "items": { "type": "string", "minLength": 1, "maxLength": 80 },
              },
            "minAge": { "type": "number", "minimum": 18, "maximum": 100 },
            "maxAge": { "type": "number", "minimum": 18, "maximum": 100 },
            "status":
              { "type": "string", "enum": ["available", "unavailable"] },
            "listing":
              {
                "type": "string",
                "enum": ["free", "weekly", "monthly", "quarterly"],
              },
            "partnerOptIn": { "type": "boolean" },
            "accurate": { "type": "boolean" },
            "terms": { "type": "boolean" },
            "adult": { "type": "boolean" },
            "photos":
              {
                "type": "array",
                "minItems": 0,
                "maxItems": 8,
                "items": { "$ref": "#/components/schemas/RegistrationPhoto" },
              },
            "mainPhoto": { "type": "number", "minimum": 0, "maximum": 7 },
          },
        "required":
          [
            "displayName",
            "dateOfBirth",
            "email",
            "mobile",
            "password",
            "confirmPassword",
            "headline",
            "about",
            "country",
            "state",
            "city",
            "area",
            "age",
            "pinCode",
            "interests",
            "languages",
            "preferences",
            "availability",
            "minAge",
            "maxAge",
            "status",
            "listing",
            "partnerOptIn",
            "accurate",
            "terms",
            "adult",
            "photos",
            "mainPhoto",
          ],
      }
    RegistrationRecord:
      {
        "type": "object",
        "properties":
          {
            "displayName":
              { "type": "string", "minLength": 1, "maxLength": 80 },
            "dateOfBirth":
              { "type": "string", "minLength": 1, "maxLength": 10 },
            "email": { "type": "string", "minLength": 1, "maxLength": 254 },
            "mobile": { "type": "string", "minLength": 1, "maxLength": 25 },
            "headline": { "type": "string", "minLength": 0, "maxLength": 120 },
            "about": { "type": "string", "minLength": 0, "maxLength": 2000 },
            "country": { "type": "string", "minLength": 0, "maxLength": 80 },
            "state": { "type": "string", "minLength": 0, "maxLength": 80 },
            "city": { "type": "string", "minLength": 0, "maxLength": 80 },
            "area": { "type": "string", "minLength": 0, "maxLength": 100 },
            "age": { "type": "number", "minimum": 18, "maximum": 100 },
            "pinCode": { "type": "string", "minLength": 0, "maxLength": 12 },
            "interests":
              {
                "type": "array",
                "minItems": 0,
                "maxItems": 20,
                "uniqueItems": true,
                "items": { "type": "string", "minLength": 1, "maxLength": 80 },
              },
            "languages":
              {
                "type": "array",
                "minItems": 0,
                "maxItems": 20,
                "uniqueItems": true,
                "items": { "type": "string", "minLength": 1, "maxLength": 80 },
              },
            "preferences":
              {
                "type": "array",
                "minItems": 0,
                "maxItems": 20,
                "uniqueItems": true,
                "items": { "type": "string", "minLength": 1, "maxLength": 80 },
              },
            "availability":
              {
                "type": "array",
                "minItems": 0,
                "maxItems": 20,
                "uniqueItems": true,
                "items": { "type": "string", "minLength": 1, "maxLength": 80 },
              },
            "minAge": { "type": "number", "minimum": 18, "maximum": 100 },
            "maxAge": { "type": "number", "minimum": 18, "maximum": 100 },
            "status":
              { "type": "string", "enum": ["available", "unavailable"] },
            "listing":
              {
                "type": "string",
                "enum": ["free", "weekly", "monthly", "quarterly"],
              },
            "partnerOptIn": { "type": "boolean" },
            "accurate": { "type": "boolean" },
            "terms": { "type": "boolean" },
            "adult": { "type": "boolean" },
            "photos":
              {
                "type": "array",
                "minItems": 0,
                "maxItems": 8,
                "items": { "$ref": "#/components/schemas/RegistrationPhoto" },
              },
            "mainPhoto": { "type": "number", "minimum": 0, "maximum": 7 },
            "id": { "type": "string", "minLength": 1, "maxLength": 120 },
            "submittedAt":
              { "type": "string", "minLength": 1, "maxLength": 120 },
            "reviewStatus":
              { "type": "string", "minLength": 1, "maxLength": 120 },
            "listingPrice": { "type": "number", "minimum": 0 },
            "isComplete": { "type": "boolean" },
            "isPublished": { "type": "boolean" },
            "missingSteps": { "type": "array", "items": { "type": "string" } },
            "notificationStatus": { "type": "string" },
            "publicSlug": { "type": "string" },
          },
        "required":
          [
            "displayName",
            "dateOfBirth",
            "email",
            "mobile",
            "headline",
            "about",
            "country",
            "state",
            "city",
            "area",
            "age",
            "pinCode",
            "interests",
            "languages",
            "preferences",
            "availability",
            "minAge",
            "maxAge",
            "status",
            "listing",
            "partnerOptIn",
            "accurate",
            "terms",
            "adult",
            "photos",
            "mainPhoto",
            "id",
            "submittedAt",
            "reviewStatus",
            "listingPrice",
          ],
      }
...