# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Him For You discovery and membership API
 * OpenAPI spec version: 0.1.0
 */
import type { RegistrationInputListing } from './registrationInputListing';
import type { RegistrationInputStatus } from './registrationInputStatus';
import type { RegistrationPhoto } from './registrationPhoto';

export interface RegistrationInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  displayName: string;
  /**
     * @minLength 1
     * @maxLength 10
     */
  dateOfBirth: string;
  /**
     * @minLength 1
     * @maxLength 254
     */
  email: string;
  /**
     * @minLength 1
     * @maxLength 25
     */
  mobile: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
  /**
     * @minLength 1
     * @maxLength 128
     */
  confirmPassword: string;
  /**
     * @minLength 0
     * @maxLength 120
     */
  headline: string;
  /**
     * @minLength 0
     * @maxLength 2000
     */
  about: string;
  /**
     * @minLength 0
     * @maxLength 80
     */
  country: string;
  /**
     * @minLength 0
     * @maxLength 80
     */
  state: string;
  /**
     * @minLength 0
     * @maxLength 80
     */
  city: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  area: string;
  /**
     * @minimum 18
     * @maximum 100
     */
  age: number;
  /**
     * @minLength 0
     * @maxLength 12
     */
  pinCode: string;
  /**
     * @minItems 0
     * @maxItems 20
     * @items.minLength 1
     * @items.maxLength 80
     */
  interests: string[];
  /**
     * @minItems 0
     * @maxItems 20
     * @items.minLength 1
     * @items.maxLength 80
     */
  languages: string[];
  /**
     * @minItems 0
     * @maxItems 20
     * @items.minLength 1
     * @items.maxLength 80
     */
  preferences: string[];
  /**
     * @minItems 0
     * @maxItems 20
     * @items.minLength 1
     * @items.maxLength 80
     */
  availability: string[];
  /**
     * @minimum 18
     * @maximum 100
     */
  minAge: number;
  /**
     * @minimum 18
     * @maximum 100
     */
  maxAge: number;
  status: RegistrationInputStatus;
  listing: RegistrationInputListing;
  partnerOptIn: boolean;
  accurate: boolean;
  terms: boolean;
  adult: boolean;
  /**
     * @minItems 0
     * @maxItems 8
     */
  photos: RegistrationPhoto[];
  /**
     * @minimum 0
     * @maximum 7
     */
  mainPhoto: number;
}