# lib/api-zod/src/generated/types/registrationPhoto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Him For You discovery and membership API
 * OpenAPI spec version: 0.1.0
 */
import type { RegistrationPhotoCategory } from './registrationPhotoCategory';

export interface RegistrationPhoto {
  /**
     * @minLength 1
     * @maxLength 1500000
     */
  dataUrl: string;
  category: RegistrationPhotoCategory;
}